# apps/web/openapi.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tivmark API",
    "version": "1.0.0",
    "description": "The API-first contract used by Tivmark and approved headless clients.",
    "license": { "name": "Proprietary", "url": "https://tivmark.com/terms" }
  },
  "servers": [
    { "url": "https://app.tivmark.com", "description": "Production" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "OAuth access token or Tivmark service key"
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/oauth/authorize",
            "tokenUrl": "/oauth/token",
            "refreshUrl": "/oauth/token",
            "scopes": {
              "profile": "Read and update the current profile",
              "teams": "Read and update teams",
              "members": "Read and manage team members",
              "invitations": "Read and manage invitations",
              "time_off": "Read and manage time-off requests",
              "time_off.approve": "Approve or decline time-off requests",
              "time_off.policy": "Manage time-off allowances",
              "credentials": "Manage service credentials and OAuth clients",
              "sso": "Manage single sign-on",
              "directory_sync": "Manage directory synchronization",
              "webhooks": "Manage outbound webhooks",
              "audit_logs": "Read audit logs",
              "billing": "Read and manage billing"
            }
          }
        }
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "image": { "type": ["string", "null"] }
        },
        "required": ["id", "name", "email", "image"]
      },
      "Problem": {
        "type": "object",
        "properties": {
          "type": { "type": "string" },
          "code": { "type": "string" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": { "type": "string" }
            }
          }
        },
        "required": ["type", "code", "title", "status", "detail"]
      },
...
      "TimeOffRequest": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "type": {
            "type": "string",
            "enum": ["VACATION", "SICK", "PERSONAL", "UNPAID"]
          },
          "status": {
            "type": "string",
            "enum": ["PENDING", "APPROVED", "DECLINED", "CANCELED"]
          },
          "startDate": { "type": "string", "format": "date" },
          "endDate": { "type": "string", "format": "date" },
          "duration": { "type": "string", "enum": ["FULL_DAY", "HALF_DAY"] },
          "halfDayPeriod": {
            "type": ["string", "null"],
            "enum": ["MORNING", "AFTERNOON"]
          },
          "requestedHalfDays": { "type": "integer" },
          "reason": { "type": ["string", "null"] },
          "reviewNote": { "type": ["string", "null"] },
          "requester": { "$ref": "#/components/schemas/User" },
          "reviewer": {
            "allOf": [
              { "$ref": "#/components/schemas/User" },
              { "type": ["object", "null"] }
            ]
          }
        },
        "required": [
          "id",
          "type",
          "status",
          "startDate",
          "endDate",
          "duration",
          "halfDayPeriod",
          "requestedHalfDays",
          "reason",
          "reviewNote",
          "requester",
          "reviewer"
        ]
      },
      "TimeOffPolicy": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "type": {
            "type": "string",
            "enum": ["VACATION", "SICK", "PERSONAL", "UNPAID"]
          },
          "annualAllowanceHalfDays": { "type": ["integer", "null"] }
        },
        "required": ["id", "type", "annualAllowanceHalfDays"]
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/teams/{teamId}/time-off/requests": {
      "get": {
        "tags": ["Time Off"],
        "summary": "List time-off requests",
        "security": [{ "oauth2": [] }, { "bearerAuth": [] }],
        "parameters": [
          {
            "schema": { "type": "string", "format": "uuid" },
            "required": true,
            "in": "path",
            "name": "teamId"
          },
          {
            "schema": { "type": ["integer", "null"] },
            "required": false,
            "name": "year",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": ["PENDING", "APPROVED", "DECLINED", "CANCELED"]
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": ["VACATION", "SICK", "PERSONAL", "UNPAID"]
            },
            "required": false,
            "name": "type",
            "in": "query"
          },
          {
            "schema": { "type": "string", "format": "uuid" },
            "required": false,
            "name": "requesterId",
            "in": "query"
          },
          {
            "schema": { "type": "string" },
            "required": false,
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": { "type": ["integer", "null"], "maximum": 100 },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Time-off requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/TimeOffRequest" }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "nextCursor": { "type": ["string", "null"] }
                      },
                      "required": ["nextCursor"]
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "409": {
            "description": "Resource conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "operationId": "get_teams_teamId_time_off_requests"
      },
      "post": {
        "tags": ["Time Off"],
        "summary": "Request time off",
        "security": [{ "oauth2": [] }, { "bearerAuth": [] }],
        "parameters": [
          {
            "schema": { "type": "string", "format": "uuid" },
            "required": true,
            "in": "path",
            "name": "teamId"
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Unique retry key. Reusing it with the same body returns the original response for 24 hours.",
            "schema": { "type": "string", "minLength": 8, "maxLength": 200 }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "requesterId": { "type": "string", "format": "uuid" },
                  "type": {
                    "type": "string",
                    "enum": ["VACATION", "SICK", "PERSONAL", "UNPAID"]
                  },
                  "startDate": { "type": "string", "format": "date" },
                  "endDate": { "type": "string", "format": "date" },
                  "duration": {
                    "type": "string",
                    "enum": ["FULL_DAY", "HALF_DAY"]
                  },
                  "halfDayPeriod": {
                    "type": ["string", "null"],
                    "enum": ["MORNING", "AFTERNOON"]
                  },
                  "reason": { "type": ["string", "null"], "maxLength": 500 }
                },
                "required": ["type", "startDate", "endDate", "duration"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/TimeOffRequest" }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "409": {
            "description": "Resource conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "operationId": "post_teams_teamId_time_off_requests"
      }
    },
...
    "/api/v1/teams/{teamId}/time-off/policies": {
      "get": {
        "tags": ["Time Off"],
        "summary": "List allowance policies",
        "security": [{ "oauth2": [] }, { "bearerAuth": [] }],
        "parameters": [
          {
            "schema": { "type": "string", "format": "uuid" },
            "required": true,
            "in": "path",
            "name": "teamId"
          }
        ],
        "responses": {
          "200": {
            "description": "Policies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/TimeOffPolicy" }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "nextCursor": { "type": ["string", "null"] }
                      },
                      "required": ["nextCursor"]
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "409": {
            "description": "Resource conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "operationId": "get_teams_teamId_time_off_policies"
      },
      "patch": {
        "tags": ["Time Off"],
        "summary": "Update allowance policies",
        "security": [{ "oauth2": [] }, { "bearerAuth": [] }],
        "parameters": [
          {
            "schema": { "type": "string", "format": "uuid" },
            "required": true,
            "in": "path",
            "name": "teamId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "allowances": {
                    "type": "object",
                    "properties": {
                      "VACATION": { "type": ["integer", "null"] },
                      "SICK": { "type": ["integer", "null"] },
                      "PERSONAL": { "type": ["integer", "null"] },
                      "UNPAID": { "type": ["integer", "null"] }
                    }
                  }
                },
                "required": ["allowances"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Policies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/TimeOffPolicy" }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "nextCursor": { "type": ["string", "null"] }
                      },
                      "required": ["nextCursor"]
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "409": {
            "description": "Resource conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "operationId": "patch_teams_teamId_time_off_policies"
      }
    },
...
  },
...
}